Attempting to reorganize the cwru_ros_pkg into a more sensical arrangement.
This first commit breaks up the stacks and reorganizes them into functional
divisions and moves all launch and config files into /config and /launch directories
in the packages.

All wheelchair driver packages are now in cwru_common. All navigation, planning, and
steering packages are now in cwru_planning.

Next steps will be to create unary stacks for all packages not in stacks and create
a stack for robot configs with a package for each robot.