Realligned the arm.
This time with more linear algebra.
The result is plausible, and looks good for some arm positions,
but not for others. For the 18 points it's calculated from, it has
RMS position error of 1.4 cm. However, the misallignment of the
meshes is obvious in rviz. I suspect the Kinect is nonlinear and
inaccurate.