Pick and place nominally working for low table.
With a low table, the pick and place sort of works.
The kinect is still super innacurate, and I think it's probably
non-linear, so I'll have to calibrate it to be right-ish at table height,
since that's the only place it matters.
Collision checking doesn't always seem to work/run/be a thing. Not sure
why.

Next: scripted task.